+++ CHANGELOG.md
@@ -2,7 +2,6 @@
 
 ## WorkOn
 
-- WorkOn: Using state management to manage user selected contest and problems
 - WorkOn: Disable puppeteer restore pages notification when app started
 - WorkOn: Complete problem operation API (e.g.: Submit answer)
 - WorkOn: Complete account API and exposers
@@ -11,6 +10,7 @@
 ## 0.1.4
 
 - Update: Use `immer` middleware in contest state management
+- Update: Using state management to manage user selected contest and problems
 
 ## 0.1.3